Clarify what reply_code should be used for
Similar to ROS 1 services, it encodes whether the server was able to handle the service call. NOT whether the service handler was successful (ie: it doesn't reflect whether the action which was performed as a result of invoking the service was successful, but whether the server is capable of handling the requests, which are two different things).